+++ src/Lykke.SettingsReader/SettingsTemplate/DefaultJsonTemplateGenerator.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Collections.Generic;
 using System.Text.Json;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
 
 namespace Lykke.SettingsReader.SettingsTemplate;
 /// <summary>
@@ -8,18 +10,22 @@ namespace Lykke.SettingsReader.SettingsTemplate;
 /// </summary>
 internal class DefaultJsonTemplateGenerator : IJsonSettingsTemplateGenerator
 {
+    private readonly ILoggerFactory  _loggerFactory;
     private readonly IConfiguration _configuration;
+    private readonly TemplateFilers _regexFilters;
 
-    public DefaultJsonTemplateGenerator(IConfiguration configuration)
+    public DefaultJsonTemplateGenerator(ILoggerFactory  loggerFactory,IConfiguration configuration, TemplateFilers regexFilters)
     {
+        _loggerFactory = loggerFactory;
         _configuration = configuration  ?? throw new ArgumentNullException(nameof(configuration));
+        _regexFilters = regexFilters;
     }
 
     public string GenerateJsonSettingsTemplate()
     {
         var json = JsonSerializer.Serialize(_configuration, new JsonSerializerOptions
         {
-            Converters = { new SettingsTemplateConverter() },
+            Converters = { new SettingsTemplateConverter(_loggerFactory,_regexFilters) },
             WriteIndented = true
         });